requirements.txt
streamlit==1.25.0
faiss-cpu==1.7.2
openai==0.27.0
langchain==0.0.314
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pinecone-client==2.2.1
pickle-mixin==1.0.2
langchain_community
